requirements.txt
python-dotenv==1.0.0
pyyaml==6.0.1
spotipy==2.23.0
requests==2.31.0
openai==1.12.0
beautifulsoup4==4.12.3
tiktoken==0.6.0
filelock==3.13.1
selenium==4.16.0